fix(structure): changed structure to comply with DDD standards
