+++ internal/domain/repositories/dividends_repository.go
@@ -1,0 +1,6 @@
+package repositories
+
+// DividendsRepository defines the interface for getting dividends per year.
+type DividendsRepository interface {
+	ListDividendsByETF(etf string) (map[string]float64, error)
+}
